+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Add `Semigroup` and `Monoid` instances for `Strict` types (thanks to
+  @ozkutuk)
+
 # 0.0.1.0
 
 * Derive `Eq`, `Ord,`, `Show`, `Read` instances for `Strict` types